+++ hpy/devel/include/common/runtime/ctx_module.h
@@ -1,0 +1,9 @@
+#ifndef HPY_COMMON_RUNTIME_CTX_MODULE_H
+#define HPY_COMMON_RUNTIME_CTX_MODULE_H
+
+#include "hpy.h"
+#include "api.h"
+
+_HPy_HIDDEN HPy ctx_Module_Create(HPyContext ctx, HPyModuleDef *hpydef);
+
+#endif /* HPY_COMMON_RUNTIME_CTX_MODULE_H */
