+++ setup.py
@@ -15,9 +15,9 @@ setup(
                   ['hpy/universal/src/hpymodule.c',
                    'hpy/universal/src/handles.c',
                    'hpy/universal/src/ctx.c',
-                   'hpy/universal/src/ctx_module.c',
                    'hpy/universal/src/ctx_meth.c',
                    'hpy/universal/src/ctx_misc.c',
+                   'hpy/devel/src/runtime/ctx_module.c',
                    'hpy/devel/src/runtime/ctx_type.c',
                    'hpy/devel/src/runtime/argparse.c',
                   ],
